+++ zhalilov.rail/P4/inputMatrix.hpp
@@ -2,10 +2,11 @@
 #define INPUT_MATRIX_HPP
 
 #include <iosfwd>
+#include <cstddef>
 
 namespace zhalilov
 {
-  std::istream &inputMatrix(int matrix[], int rows, int cols, std::istream &input);
+  std::istream &inputMatrix(int matrix[], size_t rows, size_t cols, std::istream &input);
 }
 
 #endif
